+++ darts/explainability/shap_adapters/torch_shap_adapter.py
@@ -21,8 +21,9 @@ from darts.logging import get_logger, raise_log
 from darts.models.forecasting.pl_forecasting_module import PLForecastingModule
 from darts.models.forecasting.torch_forecasting_model import TorchForecastingModel
 from darts.typing import TimeSeriesLike
+from darts.utils.data import TorchInferenceBatch
 from darts.utils.data.tabularization import create_lagged_component_names
-from darts.utils.data.torch_datasets.utils import TorchInferenceDatasetOutput
+from darts.utils.data.torch_datasets.utils import TorchInferenceSample
 from darts.utils.historical_forecasts.optimized_historical_forecasts_torch import (
     _create_dataset_bounds,
 )
@@ -89,7 +90,7 @@ class TorchShapAdapter(ShapAdapter):
         # follow the logic of `TorchForecastingModel.predict_from_dataset()`
         # to collect samples and collate them into a sample tuple
         # collect batch of samples from the end of the dataset
-        batch: list[TorchInferenceDatasetOutput] = []
+        batch: list[TorchInferenceSample] = []
         if n_samples < len(dataset):
             # randomly sample from the dataset if in training mode
             indices = np.random.choice(len(dataset), size=n_samples, replace=False)
@@ -102,20 +103,24 @@ class TorchShapAdapter(ShapAdapter):
         # - Collate each input type separately and arrange in the same feature order as SKLearnModel X array:
         #   - lagged_target | lagged_past_covariates | lagged_future_covariates | static,
         #   where lagged_future_covariates includes both historic (-ICL to -1) and actual future (0 to OCL-1)
-        # - `batch` is a list of tuples of (past target, past cov, future past cov, historic future cov, future cov,
-        #   static cov, target series schema, pred time)
-        # - since `ShapExplainer` never performs auto-regression, we can skip the "future past cov" part
-        extract_batch_indices = [0, 1, 3, 4, 5]
+        # - since `ShapExplainer` never performs auto-regression, we can skip "future past cov"
+        extract_fields = (
+            "past_target",
+            "past_covariates",
+            "historic_future_covariates",
+            "future_covariates",
+            "static_covariates",
+        )
         arrays = [
-            np.stack([sample[idx] for sample in batch])
-            for idx in extract_batch_indices
-            if batch[0][idx] is not None
+            np.stack([getattr(sample, name) for sample in batch])
+            for name in extract_fields
+            if getattr(batch[0], name) is not None
         ]
         shap_array = np.concatenate(
             [array.reshape(array.shape[0], -1) for array in arrays],
             axis=-1,
         )
-        prediction_times = pd.Index([c[-1] for c in batch])
+        prediction_times = pd.Index([sample.pred_time for sample in batch])
         return shap_array, prediction_times
 
     def _build_feature_names(self) -> list[str]:
@@ -265,12 +270,21 @@ class TorchShapAdapter(ShapAdapter):
         # set model to eval mode to deactivate dropout layers
         pl_module.eval()
 
+        def _slice_or_none(el, sl):
+            return el[sl] if el is not None else None
+
         outputs = []
         for batch_idx, i in enumerate(range(0, num_samples, self.batch_size)):
             batch_slice = slice(i, i + self.batch_size)
-            batch = tuple(
-                x_i[batch_slice] if x_i is not None else None
-                for x_i in [x_pt, x_pc, None, x_hfc, x_fc, x_sc, None, None]
+            batch = TorchInferenceBatch(
+                past_target=x_pt[batch_slice],
+                past_covariates=_slice_or_none(x_pc, batch_slice),
+                future_past_covariates=None,
+                historic_future_covariates=_slice_or_none(x_hfc, batch_slice),
+                future_covariates=_slice_or_none(x_fc, batch_slice),
+                static_covariates=_slice_or_none(x_sc, batch_slice),
+                series_schema=None,
+                pred_time=None,
             )
 
             # output shape: (num_samples = 1, batch_size, output_chunk_length, n_targets_likelihood)