+++ darts/tests/ad/test_scorers.py
@@ -1,5 +1,4 @@
 from collections.abc import Sequence
-from itertools import product
 
 import numpy as np
 import pytest
@@ -23,6 +22,7 @@ from darts.ad.scorers import DifferenceScorer as Difference
 from darts.ad.scorers import NormScorer as Norm
 from darts.ad.scorers.scorers import NLLScorer
 from darts.models import MovingAverageFilter
+from darts.tests.parametrize_helpers import param_product
 from darts.utils.timeseries_generation import linear_timeseries
 
 list_NonFittableAnomalyScorer = [
@@ -1629,7 +1629,7 @@ class TestAnomalyDetectionScorer:
 
     @pytest.mark.parametrize(
         "model,series",
-        product(
+        param_product(
             [(KMeansScorer, {"random_state": 42}), (PyODScorer, {"model": KNN()})],
             [(train, test), (mts_train, mts_test)],
         ),
@@ -1652,7 +1652,7 @@ class TestAnomalyDetectionScorer:
 
     @pytest.mark.parametrize(
         "window,model,series",
-        product(
+        param_product(
             [2, 10, 39],
             [
                 (KMeansScorer, {"random_state": 42}),